Move a few more methods into TestBase
