+++ TeachingRecordSystem/tests/TeachingRecordSystem.SupportUi.Tests/PageTests/Alerts/RolesWithoutAlertWritePermissionDataAttribute.cs
@@ -1,4 +1,4 @@
-namespace TeachingRecordSystem.SupportUi.Tests.PageTests.Alerts.EditAlert;
+namespace TeachingRecordSystem.SupportUi.Tests.PageTests.Alerts;
 
 public class RolesWithoutAlertWritePermissionDataAttribute : RoleNamesData
 {
